Ich habe noch einen Test hinzugefügt und ein paar fehler in den Klassen behoben.

+++ Library/Interface/IAusleihen.cs
@@ -8,7 +8,7 @@ namespace Library
 {
     internal interface IAusleihen
     {
-       float ausleihenPreis();
+       double ausleihenPreis();
         int Anzahltage { get; set; }
     }
 }
